+++ Vehicle.java
@@ -5,19 +5,19 @@ public interface Vehicle {
     final ArrayList<String> possibleConditions = new ArrayList<>(Arrays.asList("new", "used", "broken"));
     final ArrayList<String> possibleCleanliness = new ArrayList<>(Arrays.asList("sparkling", "clean", "dirty"));
     void setName(String name);
-    void setSaleBonus(double saleBonus);
-    void setRepairBonus(double repairBonus);
-    void setWashBonus(double washBonus);
-    void setSalePrice(double salePrice);
+    void setSaleBonus(int saleBonus);
+    void setRepairBonus(int repairBonus);
+    void setWashBonus(int washBonus);
+    void setSalePrice(int salePrice);
     void setCondition(String condition);
     void setCleanliness(String cleanliness);
 
     String getName();
-    double getSaleBonus();
-    double getRepairBonus();
-    double getWashBonus();
-    double getCost();
-    double getSalePrice();
+    int getSaleBonus();
+    int getRepairBonus();
+    int getWashBonus();
+    int getCost();
+    int getSalePrice();
     String getCondition();
     String getCleanliness();
